refactor: move example bank to example folder
