+++ src/monopoly/examples/__init__.py
@@ -1,0 +1,3 @@
+from .example_bank import ExampleBankProcessor
+
+__all__ = ["ExampleBankProcessor"]
